+++ NEWS.md
@@ -1,6 +1,6 @@
 # rwalkr 0.2.0
 
-* Added the function `shine_melb()` to launch a shiny app. It provides two basic plots to take a glimpse at the data: one is an overlaying time series plot and the other showing a heatmap of missing values.
+* Added the function `shine_melb()` to launch a shiny app. It provides two basic plots to take a glimpse at the data: one is an overlaying time series plot and the other showing a dot plot of missing values.
 
 # rwalkr 0.1.0