+++ NEWS.md
@@ -1,3 +1,6 @@
+# rwalkr 0.2.0.9000
+* Added the function `run_melb()` to pull Melbourne pedestrian data using Socrata, which is faster than `walk_melb()`.
+
 # rwalkr 0.2.0
 
 * Added the function `shine_melb()` to launch a shiny app. It provides two basic plots to take a glimpse at the data: one is an overlaying time series plot and the other showing a dot plot of missing values.